+++ loraham_kiss_tnc.h
@@ -29,6 +29,7 @@
 #define LHKT_ERR_FORMAT        -4
 #define LHKT_ERR_UNSUPPORTED   -5
 #define LHKT_ERR_NOSPACE       -6
+#define LHKT_ERR_TX_SOCKET     -7
 
 typedef struct {
     char kiss_host[LHKT_HOST_MAX];